+++ .eslintrc.js
@@ -2,13 +2,21 @@ module.exports = {
   root: true,
   extends: ['eslint-config-light'],
   globals: {
+    app: true,
+    need: true,
+    uni: true,
+    VConsole: true,
+    wx: true,
+    emonitor: true,
+    globalVars: true,
+    globalThis: true,
+    ROUTES: true,
     getCurrentPages: true,
+    AlipayJSBridge: true,
+    $: true,
     qq: true,
     weex: true,
     plus: true,
-  },
-  rules: {
-    'vue/no-v-text-v-html-on-component': 0,
   },
   settings: {
     'import/resolver': {
@@ -23,11 +31,37 @@ module.exports = {
     },
     'import/ignore': ['node_modules'],
   },
+  ignorePatterns: ['**/wxcomponents/*'],
   parserOptions: {
-    project: 'tsconfig.json',
-    tsconfigRootDir: __dirname,
-    sourceType: 'module',
+    // project: 'tsconfig.json',
+    // tsconfigRootDir: __dirname,
+    // sourceType: 'module',
+
+    ecmaVersion: 12,
+  },
+  rules: {
+    'vue/no-v-text-v-html-on-component': 0,
+    'vue/component-name-in-template-casing': ['error', 'PascalCase', {
+      registeredComponentsOnly: true,
+      ignores: [
+        'scroll-view',
+        'transition',
+        'swiper',
+        'swiper-item',
+        'web-view',
+      ],
+    }],
+    'import/no-unresolved': [2,
+      {
+        ignore: [
+          'node_modules',
+          // '^src/component/*',
+          // '^src/logic/*',
+          // '^src/page/*',
+          '@TIP_PLATFORM_NAME',
+          '^@/*',
+        ],
+      },
+    ],
   },
 };
-
-