+++ packages/press-ui/script/docs-demo/watch-demo.js
@@ -1,9 +1,10 @@
 const { watchPackages } = require('../../src/packages/script/docs-demo/watch');
 const { moveDemo } = require('../../src/packages/script/docs-demo/helper');
+const componentConfig = require('../../config/component-config.json');
 
 
 function main() {
-  moveDemo();
+  moveDemo(null, componentConfig);
 
   watchPackages(moveDemo);
 }