+++ packages/press-ui/src/packages/locale/demo-component-name.js
@@ -1,0 +1,9 @@
+export const getComponentNameFromUrl = (url = '') => {
+  if (!url.includes('?')) {
+    const list = url.split('/');
+    return list[list.length - 1];
+  }
+  const reg = /comp=([^&?]+)/;
+  const match = url.match(reg);
+  return match?.[1] || '';
+};