+++ packages/press-ui/src/packages/press-datetime-picker/helper.d.ts
@@ -1,1 +1,4 @@
 export declare const defaultFormatter: (type: string, value: number) => number;
+export declare const getYearAndMonth: (value: number) => {year: number; month: number};
+export declare const YEAR_AND_MONTH_AND_DATE: string;
+export declare const YEAR_AND_MONTH: string;