+++ packages/press-ui/src/packages/press-datetime-picker/press-datetime-picker.vue
@@ -28,7 +28,7 @@ import { isDef } from '../common/utils/validator';
 import { pickerProps } from '../press-picker-plus/shared';
 import { defaultProps, defaultOptions } from '../common/component-handler/press-component';
 import { forceUpdate, nextTick } from '../common/vue3/adapter';
-import { defaultFormatter } from './helper';
+import { defaultFormatter, YEAR_AND_MONTH_AND_DATE, YEAR_AND_MONTH, getYearAndMonth } from './helper';
 
 
 const currentYear = new Date().getFullYear();
@@ -246,7 +246,7 @@ export default {
       const results = this.getRanges().map(({ type, range }) => {
         let values = times(range[1] - range[0] + 1, (index) => {
           const value = range[0] + index;
-          return type === 'year' ? `${value}` : padZero(value);
+          return (type === 'year' || type === YEAR_AND_MONTH) ? `${value}` : padZero(value);
         });
         if (filter) {
           values = filter(type, values, innerValue);
@@ -268,8 +268,8 @@ export default {
           },
         ];
       }
-      const { maxYear, maxDate, maxMonth, maxHour, maxMinute } = this.getBoundary('max', this.innerValue);
-      const { minYear, minDate, minMonth, minHour, minMinute } = this.getBoundary('min', this.innerValue);
+      const { maxYear, maxDate, maxMonth, maxHour, maxMinute, maxYearMonth } = this.getBoundary('max', this.innerValue);
+      const { minYear, minDate, minMonth, minHour, minMinute, minYearMonth } = this.getBoundary('min', this.innerValue);
       const result = [
         {
           type: 'year',
@@ -294,6 +294,30 @@ export default {
       ];
       if (this.type === 'date') result.splice(3, 2);
       if (this.type === 'year-month') result.splice(2, 3);
+
+      if (this.type === YEAR_AND_MONTH_AND_DATE) {
+        const tResult =  [
+          {
+            type: YEAR_AND_MONTH,
+            range: [minYearMonth, maxYearMonth],
+          },
+          {
+            type: 'day',
+            range: [minDate, maxDate],
+          },
+          {
+            type: 'hour',
+            range: [minHour, maxHour],
+          },
+          {
+            type: 'minute',
+            range: [minMinute, maxMinute],
+          },
+        ];
+
+        return tResult;
+      }
+
       return result;
     },
     correctValue(value) {
@@ -327,12 +351,14 @@ export default {
       let date = 1;
       let hour = 0;
       let minute = 0;
+
       if (type === 'max') {
         month = 12;
         date = getMonthEndDay(value.getFullYear(), value.getMonth() + 1);
         hour = 23;
         minute = 59;
       }
+
       if (value.getFullYear() === year) {
         month = boundary.getMonth() + 1;
         if (value.getMonth() + 1 === month) {
@@ -345,12 +371,16 @@ export default {
           }
         }
       }
+
+      const yearMonth = year * 12 + month;
+
       return {
         [`${type}Year`]: year,
         [`${type}Month`]: month,
         [`${type}Date`]: date,
         [`${type}Hour`]: hour,
         [`${type}Minute`]: minute,
+        [`${type}YearMonth`]: yearMonth,
       };
     },
     onCancel() {
@@ -371,14 +401,16 @@ export default {
       } else {
         const indexes = picker.getIndexes();
         const values = indexes.map((value, index) => originColumns[index].values[value]);
-        const year = getTrueValue(values[0]);
-        const month = getTrueValue(values[1]);
+
+        let year = getTrueValue(values[0]);
+        let month = getTrueValue(values[1]);
         const maxDate = getMonthEndDay(year, month);
 
         let date = getTrueValue(values[2]);
         if (this.type === 'year-month') {
           date = 1;
         }
+
 
         date = date > maxDate ? maxDate : date;
         let hour = 0;
@@ -387,6 +419,17 @@ export default {
           hour = getTrueValue(values[3]);
           minute = getTrueValue(values[4]);
         }
+
+
+        if (this.type === YEAR_AND_MONTH_AND_DATE) {
+          const info = getYearAndMonth(values[0]);
+          year = info.year;
+          month = info.month;
+          date = getTrueValue(values[1]);
+          hour = getTrueValue(values[2]);
+          minute = getTrueValue(values[3]);
+        }
+
         value = new Date(year, month - 1, date, hour, minute);
       }
 
@@ -412,15 +455,26 @@ export default {
           formatter('year', `${date.getFullYear()}`, innerValue),
           formatter('month', padZero(date.getMonth() + 1), innerValue),
         ];
+
         if (type === 'date') {
           values.push(formatter('day', padZero(date.getDate()), innerValue));
         }
+
         if (type === 'datetime') {
           values.push(
             formatter('day', padZero(date.getDate()), innerValue),
             formatter('hour', padZero(date.getHours()), innerValue),
             formatter('minute', padZero(date.getMinutes()), innerValue),
           );
+        }
+
+        if (type === YEAR_AND_MONTH_AND_DATE) {
+          values = [
+            formatter(YEAR_AND_MONTH, date.getFullYear() * 12 + date.getMonth() + 1, innerValue),
+            formatter('day', padZero(date.getDate()), innerValue),
+            formatter('hour', padZero(date.getHours()), innerValue),
+            formatter('minute', padZero(date.getMinutes()), innerValue),
+          ];
         }
       }
 
@@ -433,5 +487,4 @@ export default {
 
 </script>
 <style scoped lang="scss">
-// @import "../common/style/press/index.scss";
 </style>